meishi_backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-filter
pillow
phonenumbers
geopy       
djangorestframework
django-rest-swagger
drf-yasg
timezonefinder
celery
redis
django-celery-beat
gunicorn
whitenoise
